domparsing/serializing_xhtml_doctypevariation3.xhtml

<!DOCTYPE  HTml  PUBLIC    "  xxx test  -//TEST /Little Test  EN  "     " external.dtd " [   <!ELEMENT  myelement ( #PCDATA)>   ]  >
<html xmlns="http://www.w3.org/1999/xhtml">
 <head>
  <title>DOM Serialization in XHTML - variation 3</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
  <div id="log"></div>
  <script><![CDATA[
  test(function(){
    // Note: line endings are not interoperable (for windows-systems, some
    // browsers let literal \r\n through their XML parser, while others
    // convert \r\n to \n.

    // Set testing expectations based on platform capability
    var expectedSerialization = (typeof document.doctype.internalSubset != undefined) ?
        "<!DOCTYPE Html PUBLIC \"  xxx test  -//TEST /Little Test  EN  \" \" external.dtd \" [   <!ELEMENT  myelement ( #PCDATA)>   ]>" :
        "<!DOCTYPE Html PUBLIC \"  xxx test  -//TEST /Little Test  EN  \" \" external.dtd \">";
    assert_equals(new XMLSerializer().serializeToString(document.doctype), expectedSerialization);
  }, "Parsed DOCTYPE serialized correctly");

  // ]]>
  </script>
 </body>
</html>
